Struct rustc_privacy::NamePrivacyVisitor

pub(crate) struct NamePrivacyVisitor<'tcx> {
    pub(crate) tcx: TyCtxt<'tcx>,
    pub(crate) maybe_typeck_results: Option<&'tcx TypeckResults<'tcx>>,
}

Name privacy visitor, checks privacy and reports violations. Most of name privacy checks are performed during the main resolution phase, or later in type checking when field accesses and associated items are resolved. This pass performs remaining checks for fields in struct expressions and patterns.

Fields
tcx: TyCtxt<'tcx>maybe_typeck_results: Option<&'tcx TypeckResults<'tcx>>
Implementations

impl<'tcx> NamePrivacyVisitor<'tcx>

pub(crate) fn typeck_results(&self) -> &'tcx TypeckResults<'tcx>

Gets the type-checking results for the current body. As this will ICE if called outside bodies, only call when working with Expr or Pat nodes (they are guaranteed to be found only in bodies).

pub(crate) fn check_field( &mut self, hir_id: HirId, use_ctxt: Span, span: Span, def: AdtDef<'tcx>, field: &'tcx FieldDef, in_update_syntax: bool )