Struct rustc_privacy::TypePrivacyVisitor

pub(crate) struct TypePrivacyVisitor<'tcx> {
    pub(crate) tcx: TyCtxt<'tcx>,
    pub(crate) maybe_typeck_results: Option<&'tcx TypeckResults<'tcx>>,
    pub(crate) current_item: LocalDefId,
    pub(crate) span: Span,
}

Type privacy visitor, checks types for privacy and reports violations. Both explicitly written types and inferred types of expressions and patterns are checked. Checks are performed on “semantic” types regardless of names and their hygiene.

Fields
tcx: TyCtxt<'tcx>maybe_typeck_results: Option<&'tcx TypeckResults<'tcx>>current_item: LocalDefIdspan: Span
Implementations

impl<'tcx> TypePrivacyVisitor<'tcx>

pub(crate) fn typeck_results(&self) -> &'tcx TypeckResults<'tcx>

Gets the type-checking results for the current body. As this will ICE if called outside bodies, only call when working with Expr or Pat nodes (they are guaranteed to be found only in bodies).

pub(crate) fn item_is_accessible(&self, did: DefId) -> bool

pub(crate) fn check_expr_pat_type(&mut self, id: HirId, span: Span) -> bool

pub(crate) fn check_def_id( &mut self, def_id: DefId, kind: &str, descr: &dyn Display ) -> bool