Struct rustc_privacy::SearchInterfaceForPrivateItemsVisitor

pub(crate) struct SearchInterfaceForPrivateItemsVisitor<'tcx> {
    pub(crate) tcx: TyCtxt<'tcx>,
    pub(crate) item_def_id: LocalDefId,
    pub(crate) required_visibility: Visibility,
    pub(crate) required_effective_vis: Option<EffectiveVisibility>,
    pub(crate) in_assoc_ty: bool,
    pub(crate) in_primary_interface: bool,
}

SearchInterfaceForPrivateItemsVisitor traverses an item’s interface and finds any private components in it. PrivateItemsInPublicInterfacesVisitor ensures there are no private types and traits in public interfaces.

Fields
tcx: TyCtxt<'tcx>item_def_id: LocalDefIdrequired_visibility: Visibility

The visitor checks that each component type is at least this visible.

required_effective_vis: Option<EffectiveVisibility>in_assoc_ty: boolin_primary_interface: bool
Implementations

impl SearchInterfaceForPrivateItemsVisitor<'_>

pub(crate) fn generics(&mut self) -> &mut Self

pub(crate) fn predicates(&mut self) -> &mut Self

pub(crate) fn bounds(&mut self) -> &mut Self

pub(crate) fn ty(&mut self) -> &mut Self

pub(crate) fn check_def_id( &mut self, def_id: DefId, kind: &str, descr: &dyn Display, ) -> bool

pub(crate) fn leaks_private_dep(&self, item_id: DefId) -> bool

An item is ‘leaked’ from a private dependency if all of the following are true:

It’s contained within a public type

It comes from a private crate